+++ src/__routes/__layout.svelte
@@ -1,10 +1,8 @@
 <script>
   import "$lib/css/app.css";
-  import "$assets/css/oceanic.css";
-  import "$assets/css/markdown.css";
 
   import store from "$misc/store";
-  import icon_arrow from "$assets/icons/arrow.svg";
+  import icon_arrow from "./arrow.svg";
 
   import { goto } from "$app/navigation";
   import Sidebar from "$lib/components/sidebar/Sidebar.svelte";
