+++ src/__routes/examples/auth/index.svelte
