+++ src/__routes/examples/file/index.svelte
