+++ src/__routes/examples/index.svelte
