+++ src/__routes/examples/pokedex/[pokemon].svelte
