+++ src/__routes/examples/pokedex/index.svelte
