+++ src/__routes/examples/settings/index.svelte
