+++ src/assets/markdown/welcome/00-warning.md
@@ -1,0 +1,7 @@
+---
+color: error-lighter
+---
+
+## Beta
+
+Svelte Kit is in beta and there will be some API changes before they hit V1. Bubbles will not hit V1 until Svelte Kit does.
