+++ src/assets/markdown/welcome/01-intro.md
@@ -1,0 +1,23 @@
+---
+---
+
+## Usage
+
+<br>
+
+**Who should and shouldn't use Bubbles**
+
+Bubbles is an opinionated component library build for Svelte Kit. It's has a specific structure and style that it follows. If the overall structure of it's layouts and components don't fit your needs, you're most likely better of using a more flexible component framework like Material UI, Tailwind, or writing your own.
+
+Bubbles is mostly set up to construct dashboards NOT landing pages and marketing sites. It has lots of dashboard specific components like Sidebars and Tables that are usually overkill and/or are too restrictive for marketing sites.
+
+<br>
+
+**Performance**
+
+Bubbles is build specifically for Svelte Kit. This means that it installs as a dev dependency and will not
+include any components you don't use after your build step.
+
+This one-page knowledgebase currently imports all of the components into one page - something that is not
+realistic of the real world. These docs could have some minor performances issues while you're using them
+(specifically when you're resizing the window).
